+++ frontend/.storybook/main.js
@@ -1,37 +1,27 @@
 const { TsconfigPathsPlugin } = require('tsconfig-paths-webpack-plugin');
-
 module.exports = {
   stories: ['../**/*.stories.mdx', '../**/*.stories.@(js|jsx|ts|tsx)'],
   addons: [
-    '@storybook/preset-scss',
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/addon-a11y',
-    'storybook-addon-next-router',
     '@chakra-ui/storybook-addon',
-    // 'storybook-addon-performance/register',
-    {
-      name: '@storybook/addon-postcss',
-      options: {
-        postcssLoaderOptions: {
-          implementation: require('postcss'),
-        },
-      },
-    },
+    '@storybook/addon-styling',
   ],
-  core: {
-    builder: 'webpack5',
+  framework: {
+    name: '@storybook/nextjs',
+    options: {
+      fastRefresh: true,
+    },
   },
-  framework: '@storybook/react',
   staticDirs: ['../public'],
-  reactOptions: {
-    fastRefresh: true,
-  },
   features: {
     emotionAlias: false,
     postcss: false,
   },
-  typescript: { reactDocgen: false },
+  typescript: {
+    reactDocgen: false,
+  },
   webpackFinal: async (config) => {
     config.resolve.plugins = [
       ...(config.resolve.plugins || []),
@@ -40,5 +30,8 @@ module.exports = {
       }),
     ];
     return config;
+  },
+  docs: {
+    autodocs: true,
   },
 };